Notificacao Moradas
é possivel notificar o utilizador de coisas perto das suas moradas

+++ FinalProject/app/src/main/java/com/example/finalproject/firebase/dao/AddressDao.kt
@@ -11,7 +11,7 @@ class AddressDao {
                     FieldValue.arrayUnion(hashMapOf(
                             "Address" to address,
                             "Description" to description,
-                            "coordinates" to hashMapOf("lat" to lat, "lon" to lon)
+                            "coordinates" to hashMapOf("lat" to lat.toDouble(), "lon" to lon.toDouble())
                     )))
         }
